@@ -2,14 +2,14 @@
   "name": "hikarujs",
   "version": "1.13.4",
   "description": "A static site generator that generates routes based on directories naturally.",
-  "main": "hikaru/index.js",
+  "type": "module",
   "bin": {
-    "hikaru": "bin/hikaru"
+    "hikaru": "bin/hikaru.js"
   },
   "scripts": {
     "build:docs": "jsdoc -c .jsdoc.json -R README.md -u tutorials/ && echo \"hikaru.alynx.one\" > docs/CNAME && touch docs/.nojekyll && git add -A docs/ && git commit -m \"Updated docs.\"",
-    "pref": "0x -o bin/hikaru build",
-    "test": "standardx bin/* hikaru/*.js tests/*.js && mocha tests/index.js",
+    "pref": "0x -o bin/hikaru.js build",
+    "test": "standardx bin/*.js hikaru/*.js tests/*.js && mocha tests/index.js",
     "preversion": "npm test",
     "version": "npm run build:docs"
   },
@@ -30,25 +30,25 @@
   "homepage": "https://hikaru.alynx.one/",
   "dependencies": {
     "chokidar": "^3.5.3",
-    "commander": "^9.0.0",
-    "fs-extra": "^10.0.0",
-    "highlight.js": "^11.5.0",
+    "commander": "^9.4.1",
+    "fs-extra": "^11.1.0",
+    "highlight.js": "^11.7.0",
     "is-glob": "^4.0.3",
-    "isbinaryfile": "^4.0.8",
-    "marked": "^4.0.12",
+    "isbinaryfile": "^5.0.0",
+    "marked": "^4.2.4",
     "nunjucks": "^3.2.3",
-    "parse5": "^6.0.1",
+    "parse5": "^7.1.2",
     "picomatch": "^2.3.1",
-    "yaml": "^1.10.2"
+    "yaml": "^2.1.3"
   },
   "devDependencies": {
-    "0x": "^5.1.2",
-    "chai": "^4.3.6",
-    "jsdoc": "^3.6.10",
-    "mocha": "^9.2.0",
+    "0x": "^5.4.1",
+    "chai": "^4.3.7",
+    "jsdoc": "^4.0.0",
+    "mocha": "^10.2.0",
     "standardx": "^7.0.0"
   },
   "engines": {
-    "node": ">=12.20.0"
+    "node": ">=14.16.0"
   }
 }
